+++ app/src/lib/archive-engine/constants.ts
@@ -4,5 +4,6 @@
 export enum EVENTS {
     ARCHIVE_START = `archive_start`, // Path
     PERSISTING_START = `persist_start`, // Number of assets
+    REMOTE_DELETE = `remote_delete`, // Number of remotely deleted assets
     ARCHIVE_DONE = `archive_done`
 }